requirements.txt
black==23.3.0
discord==1.7.3
humanize==4.3.0
pytest==7.2.2
python_dotenv==1.0.1
